Fix shadcn component imports and add app tests
