+++ frontend/src/App.test.tsx
@@ -1,0 +1,28 @@
+import { fireEvent, render, screen, waitFor } from '@testing-library/react'
+
+import App from './App'
+
+describe('App integration', () => {
+  it('renders shadcn sidebar components alongside the sampler UI', () => {
+    render(<App />)
+
+    expect(screen.getByText('Acme Inc.')).toBeInTheDocument()
+    expect(screen.getByText('Documents')).toBeInTheDocument()
+    expect(screen.getByText('Transport')).toBeInTheDocument()
+  })
+
+  it('allows collapsing the sidebar with the trigger control', async () => {
+    const { container } = render(<App />)
+
+    const sidebar = container.querySelector('[data-slot="sidebar"]')
+    expect(sidebar).not.toBeNull()
+    expect(sidebar).toHaveAttribute('data-state', 'expanded')
+
+    const trigger = screen.getByRole('button', { name: /toggle sidebar/i })
+    fireEvent.click(trigger)
+
+    await waitFor(() =>
+      expect(sidebar).toHaveAttribute('data-state', 'collapsed'),
+    )
+  })
+})
