+++ frontend/src/components/chart-area-interactive.tsx
