+++ frontend/src/components/section-cards.tsx
