+++ frontend/src/components/ui/avatar.tsx
