+++ frontend/src/test/setup.ts
@@ -1,6 +1,25 @@
 import '@testing-library/jest-dom';
 import { vi } from 'vitest';
 
+Object.defineProperty(window, 'innerWidth', {
+  writable: true,
+  value: 1024,
+});
+
+Object.defineProperty(window, 'matchMedia', {
+  writable: true,
+  value: vi.fn().mockImplementation((query: string) => ({
+    matches: false,
+    media: query,
+    onchange: null,
+    addListener: vi.fn(),
+    removeListener: vi.fn(),
+    addEventListener: vi.fn(),
+    removeEventListener: vi.fn(),
+    dispatchEvent: vi.fn(),
+  })),
+});
+
 // Mock ResizeObserver
 const ResizeObserverMock = vi.fn(() => ({
   observe: vi.fn(),
